Refactoring Effect (again)
* Removing `shoes/effect{.c,.h}` in favour of `shoes/types/effect`.
* Shoes3/shoes3#333
