--- shoes/effects.h
@@ -1,7 +1,0 @@
-//
-// shoes/effects.h
-// The presently meager effects API.
-//
-void shoes_gaussian_blur_filter(cairo_t *, VALUE, shoes_place *);
-void shoes_shadow_filter(cairo_t *, VALUE, shoes_place *);
-void shoes_glow_filter(cairo_t *, VALUE, shoes_place *);
